refactor: add links to translation
